Add truncate method to question history
Added a new truncate method to the question history repository, allowing history data to be cleared with a single request. This method has been added to both the actual repository and its mocks. As a result, we've also included an endpoint in the HTTP controller to trigger this operation. Alongside this, the testing infrastructure has been updated to make use of this new functionality.